feat: correct ID type in EquipmentRepository (#3)

+++ src/main/java/com/newfit/reservation/repository/EquipmentRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EquipmentRepository extends JpaRepository<Equipment, Integer> {
+public interface EquipmentRepository extends JpaRepository<Equipment, Long> {
 }
